Handle multiple login methods & invalid login data
