+++ tests/cypress/support/index.js
@@ -1,1 +1,2 @@
 import './commands';
+import './iframe-commands';
